requirements.txt
Django==1.5.1
PyYAML==3.10
argparse==1.2.1
defusedxml==0.4.1
django-guardian==1.1.1
django-tastypie==0.10.0
lxml==3.2.3
mimeparse==0.1.3
python-dateutil==2.1
requests==1.2.3
six==1.3.0
wsgiref==0.1.2
